update each servo one at a time
Change the code to update the servos one at a time, each time roll the
working cb list into the active one. @todo It might be desirable to have
a batch update routine also since the servos might need to be synced.

Also updated the unit test to test this (note the first up() works fine
but the down() is broken)